+++ saleor/attribute/utils.py
@@ -1,10 +1,11 @@
 from collections.abc import Iterable
 from typing import TYPE_CHECKING, Union
 
+from django.db.models.expressions import Exists, OuterRef
+
 from ..page.models import Page
 from ..product.models import Product, ProductVariant
 from .models import (
-    AssignedPageAttribute,
     AssignedPageAttributeValue,
     AssignedProductAttribute,
     AssignedProductAttributeValue,
@@ -14,21 +15,36 @@ from .models import (
     AttributeValue,
 )
 
-AttributeAssignmentType = Union[
-    AssignedProductAttribute, AssignedVariantAttribute, AssignedPageAttribute
-]
+AttributeAssignmentType = Union[AssignedVariantAttribute, AssignedProductAttribute]
 T_INSTANCE = Union[Product, ProductVariant, Page]
-
 
 if TYPE_CHECKING:
     from .models import AttributePage, AttributeProduct, AttributeVariant
+
+
+def disassociate_attributes_from_instance(
+    instance: T_INSTANCE,
+    *attributes: Attribute,
+) -> None:
+    """Remove attribute assigned to an instance.
+
+    This has to remove a FK to Attribute from Page instance and
+    remove all value assignments related to that same product and attribute.
+    """
+    values = AttributeValue.objects.filter(
+        attribute_id__in=[attr.id for attr in attributes]
+    )
+    AssignedPageAttributeValue.objects.filter(
+        Exists(values.filter(id=OuterRef("value_id"))),
+        page_id=instance.pk,
+    ).delete()
 
 
 def associate_attribute_values_to_instance(
     instance: T_INSTANCE,
     attribute: Attribute,
     *values: AttributeValue,
-) -> AttributeAssignmentType:
+) -> Union[None, AttributeAssignmentType]:
     """Assign given attribute values to a product or variant.
 
     Note: be aware this function invokes the ``set`` method on the instance's
@@ -41,27 +57,7 @@ def associate_attribute_values_to_instance(
     validate_attribute_owns_values(attribute, values_ids)
 
     # Associate the attribute and the passed values
-    assignment = _associate_attribute_to_instance(instance, attribute.pk)
-    assignment.values.set(values)
-
-    # While migrating to a new structure we need to make sure we also
-    # copy the assigned product to AssignedProductAttributeValue
-    # where it will live after issue #12881 will be implemented
-    if isinstance(instance, Product):
-        AssignedProductAttributeValue.objects.filter(
-            assignment_id=assignment.pk
-        ).update(product_id=instance.pk)
-
-    # This code will be deleted in new release (3.17), it is temporary solution between
-    # releases to keep database in sync
-    elif isinstance(instance, Page):
-        AssignedPageAttributeValue.objects.filter(assignment_id=assignment.pk).update(
-            page_id=instance.pk
-        )
-
-    sort_assigned_attribute_values(instance, assignment, values)
-
-    return assignment
+    return _associate_attribute_to_instance(instance, attribute, *values)
 
 
 def validate_attribute_owns_values(attribute: Attribute, value_ids: set[int]) -> None:
@@ -79,48 +75,85 @@ def validate_attribute_owns_values(attribute: Attribute, value_ids: set[int]) ->
 
 
 def _associate_attribute_to_instance(
-    instance: T_INSTANCE, attribute_pk: int
-) -> AttributeAssignmentType:
-    """Associate a given attribute to an instance."""
+    instance: T_INSTANCE,
+    attribute: Attribute,
+    *values: AttributeValue,
+) -> Union[None, AttributeAssignmentType]:
+    """Associate a given attribute to an instance.
+
+    For a given instance assign an attribute to it and set values based on *values.
+
+    Note: this will clean any value that already exist there.
+
+    This function is under rebuilding while we move away from intermediate models
+    for attribute relations
+
+    See:
+    https://github.com/saleor/saleor/issues/12881
+    """
+    if isinstance(instance, Page):
+        # Clear all assignments that don't match the values we'd like to assign
+        values_qs = AttributeValue.objects.filter(attribute_id=attribute.pk)
+        AssignedPageAttributeValue.objects.filter(
+            Exists(values_qs.filter(id=OuterRef("value_id"))),
+            page_id=instance.pk,
+        ).exclude(value__in=values).delete()
+
+        # Create new assignments
+        for value in values:
+            AssignedPageAttributeValue.objects.get_or_create(page=instance, value=value)
+
+        sort_assigned_attribute_values(instance, attribute, values)
+        return None
+
     assignment: AttributeAssignmentType
+
     if isinstance(instance, Product):
         attribute_rel: Union[
             "AttributeProduct", "AttributeVariant", "AttributePage"
-        ] = instance.product_type.attributeproduct.get(attribute_id=attribute_pk)
+        ] = instance.product_type.attributeproduct.get(attribute_id=attribute.pk)
 
         assignment, _ = AssignedProductAttribute.objects.get_or_create(
             product=instance, assignment=attribute_rel
         )
-    elif isinstance(instance, ProductVariant):
-        attribute_rel = instance.product.product_type.attributevariant.get(
-            attribute_id=attribute_pk
+        assignment.values.set(values)
+
+        sort_assigned_attribute_values_using_assignment(instance, assignment, values)
+
+        # While migrating to a new structure we need to make sure we also
+        # copy the assigned product to AssignedProductAttributeValue
+        # where it will live after issue #12881 will be implemented
+        AssignedProductAttributeValue.objects.filter(
+            assignment_id=assignment.pk
+        ).update(product_id=instance.pk)
+
+        return assignment
+
+    if isinstance(instance, ProductVariant):
+        attribute_variant = instance.product.product_type.attributevariant.get(
+            attribute_id=attribute.pk
         )
 
         assignment, _ = AssignedVariantAttribute.objects.get_or_create(
-            variant=instance, assignment=attribute_rel
+            variant=instance, assignment=attribute_variant
         )
-    elif isinstance(instance, Page):
-        attribute_rel = instance.page_type.attributepage.get(attribute_id=attribute_pk)
-        assignment, _ = AssignedPageAttribute.objects.get_or_create(
-            page=instance, assignment=attribute_rel
-        )
-    else:
-        raise AssertionError(f"{instance.__class__.__name__} is unsupported")
+        assignment.values.set(values)
 
-    return assignment
+        sort_assigned_attribute_values_using_assignment(instance, assignment, values)
+        return assignment
+
+    raise AssertionError(f"{instance.__class__.__name__} is unsupported")
 
 
-def sort_assigned_attribute_values(
+def sort_assigned_attribute_values_using_assignment(
     instance: T_INSTANCE,
     assignment: AttributeAssignmentType,
     values: Iterable[AttributeValue],
 ) -> None:
     """Sort assigned attribute values based on values list order."""
-
     instance_to_value_assignment_mapping = {
-        "Product": ("productvalueassignment", AssignedProductAttributeValue),
         "ProductVariant": ("variantvalueassignment", AssignedVariantAttributeValue),
-        "Page": ("pagevalueassignment", AssignedPageAttributeValue),
+        "Product": ("productvalueassignment", AssignedProductAttributeValue),
     }
     assignment_lookup, assignment_model = instance_to_value_assignment_mapping[
         instance.__class__.__name__
@@ -135,3 +168,23 @@ def sort_assigned_attribute_values(
         value_assignment.sort_order = index
 
     assignment_model._default_manager.bulk_update(values_assignment, ["sort_order"])
+
+
+def sort_assigned_attribute_values(
+    instance: Page,
+    attribute: Attribute,
+    values: Iterable[AttributeValue],
+) -> None:
+    values_pks = [value.pk for value in values]
+
+    values_qs = AttributeValue.objects.filter(attribute_id=attribute.pk)
+    values_assignment = list(
+        instance.attributevalues.filter(
+            Exists(values_qs.filter(id=OuterRef("value_id"))),
+        )
+    )
+    values_assignment.sort(key=lambda e: values_pks.index(e.value_id))
+    for index, value_assignment in enumerate(values_assignment):
+        value_assignment.sort_order = index
+
+    AssignedPageAttributeValue.objects.bulk_update(values_assignment, ["sort_order"])